// src/app/model/employee.ts
export class employeeModel {
    empId: number;
    empName: string;
    empContact: string;
    empEmail: string;
    empAddress: string;
    empCity: string;
    empState: string;
    empPincode: string;

    constructor( ) {
        this.empId = 1,
        this.empName = "",
        this.empAddress = "",
        this.empContact = "",
        this.empEmail = "",
        this.empCity = "",
        this.empState = "",
        this.empPincode = ""

    }

}